Fix query.

+++ backend/common/db.sql
@@ -1,6 +1,6 @@
 create table user_profile
 (
-    user_id        numeric primary key,
+    user_id        integer primary key,
     mini_app_token varchar(512) not null,
     access_token   varchar(512) not null
 );
